Fixed some implicit conversion warnings
The fix here is to cast the double to CGFLOAT_TYPE. This will do the
right thing regardless of whether CGFloat is defined as a 32 or 64 bit
float.

Additionally, the priority field is defined as a float on iOS and on the
Mac, so `-doubleValue` has been changed to `-floatValue` when extracting
the priority from `constraintValueNumberDict`.